// src/models/SocietyMaterialInwardEntriesModel.tsx
export interface SocietyMaterialInwardEntriesModel {
    SocietyMaterialInwardEntriesId: string;
    MaterialInwardToUserId: string;
    Company: string;
    NoOfParcels: number;
    Remarks: string;
    MaterialInwardStatus: string;
    ReceivedDateTime: Date;
    SocietyId: string;
    WelcomeCode: string;
    MaterialInwardToSocietyBuildingUnitId: string;
    MaterialInwardToUserType: string;
    MaterialInwardToUserTypeUserId: string;

}